Renamed files to CamelCase
