+++ src/JsonCommentStripper.ts
